+++ addOns/automation/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Added
+- Warning if unexpected element under `activeScan` job (e.g. misplaced `rules`).
 
 ## [0.20.0] - 2022-12-14
 ### Added